fix: model corrections for type safety and query accuracy
- Answer: wrap sectionId assignment in valueOrNull() to avoid storing 0
  for absent section (consistent with section_question_id handling)
- Relationship: add findTextByDownstream_S_ID named query and static
  finder to look up TEXT-action relationships targeting a given section;
  used by QuestionManager to re-evaluate tokens on section display
- RespondentPSA: change respondentId, psaId, tries fields from long to
  int to match schema column types and avoid type-mismatch mapping errors
- Survey: change reports and postSurveyActions fetch from LAZY to EAGER
  so collections are available outside a transaction in Vaadin views

+++ src/main/java/com/elicitsoftware/model/RespondentPSA.java
@@ -56,13 +56,13 @@ public class RespondentPSA extends PanacheEntityBase {
     public Integer id;
 
     @Column(name = "respondent_id")
-    public long respondentId;
+    public int respondentId;
 
     @Column(name = "post_survey_action_id")
-    public long psaId;
+    public int psaId;
 
     @Column(name = "tries")
-    public long tries;
+    public int tries;
 
     @Column(name = "status")
     public String status;
